% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyNimbleFun.R
\name{toyNimbleFun}
\alias{toyNimbleFun}
\title{toyNimbleFun}
\usage{
toyNimbleFun(model, node)
}
\arguments{
\item{model}{a nimbleModel}

\item{node}{a character string naming a single node in the provided nimbleModel}
}
\description{
toyNimbleFun
}
\details{
This function tests whether there's an issue with nimbleFunction namespacing
when making packages with NIMBLE. This comes up if you're trying to define
and use a NIMBLE class.
}
\author{
Ben Goldstein
}
